extracted config reducer from main reducer
