+++ src/organisms/Footer/Footer.tsx
@@ -1,12 +1,13 @@
 import * as React from 'react';
 import { useAppSelector } from '../../redux/hooks';
+import path from 'path';
 
 const Footer = () => {
-  const statusText = useAppSelector(state => state.statusText);
+  const rootEntry = useAppSelector(state => state.main.rootEntry);
 
   return (
     <div>
-      ManifestUI kubeshop.io 2021 - {statusText}
+      ManifestUI kubeshop.io 2021 {rootEntry.children && ('- ' + rootEntry.folder + path.sep + rootEntry.name)}
     </div>
   )
 };
